MM-12342: merge the experimental channel store
