+++ app/app.go
@@ -219,14 +219,6 @@ func New(options ...Option) (outApp *App, outErr error) {
 
 	app.Srv.Store = app.newStore()
 
-	app.AddConfigListener(func(_, current *model.Config) {
-		if current.SqlSettings.EnablePublicChannelsMaterialization != nil && !*current.SqlSettings.EnablePublicChannelsMaterialization {
-			app.Srv.Store.Channel().DisableExperimentalPublicChannelsMaterialization()
-		} else {
-			app.Srv.Store.Channel().EnableExperimentalPublicChannelsMaterialization()
-		}
-	})
-
 	if err := app.ensureAsymmetricSigningKey(); err != nil {
 		return nil, errors.Wrapf(err, "unable to ensure asymmetric signing key")
 	}
