+++ app/diagnostics.go
@@ -184,60 +184,60 @@ func (a *App) trackActivity() {
 		"active_users_daily":           activeUsersDailyCount,
 		"active_users_monthly":         activeUsersMonthlyCount,
 		"registered_deactivated_users": inactiveUserCount,
-		"teams":                        teamCount,
-		"public_channels":              publicChannelCount,
-		"private_channels":             privateChannelCount,
-		"direct_message_channels":      directChannelCount,
-		"public_channels_deleted":      deletedPublicChannelCount,
-		"private_channels_deleted":     deletedPrivateChannelCount,
-		"posts":                        postsCount,
+		"teams":                    teamCount,
+		"public_channels":          publicChannelCount,
+		"private_channels":         privateChannelCount,
+		"direct_message_channels":  directChannelCount,
+		"public_channels_deleted":  deletedPublicChannelCount,
+		"private_channels_deleted": deletedPrivateChannelCount,
+		"posts":                    postsCount,
 	})
 }
 
 func (a *App) trackConfig() {
 	cfg := a.Config()
 	a.SendDiagnostic(TRACK_CONFIG_SERVICE, map[string]interface{}{
-		"web_server_mode":                                         *cfg.ServiceSettings.WebserverMode,
-		"enable_security_fix_alert":                               *cfg.ServiceSettings.EnableSecurityFixAlert,
-		"enable_insecure_outgoing_connections":                    *cfg.ServiceSettings.EnableInsecureOutgoingConnections,
-		"enable_incoming_webhooks":                                cfg.ServiceSettings.EnableIncomingWebhooks,
-		"enable_outgoing_webhooks":                                cfg.ServiceSettings.EnableOutgoingWebhooks,
-		"enable_commands":                                         *cfg.ServiceSettings.EnableCommands,
-		"enable_only_admin_integrations":                          *cfg.ServiceSettings.EnableOnlyAdminIntegrations,
-		"enable_post_username_override":                           cfg.ServiceSettings.EnablePostUsernameOverride,
-		"enable_post_icon_override":                               cfg.ServiceSettings.EnablePostIconOverride,
-		"enable_user_access_tokens":                               *cfg.ServiceSettings.EnableUserAccessTokens,
-		"enable_custom_emoji":                                     *cfg.ServiceSettings.EnableCustomEmoji,
-		"enable_emoji_picker":                                     *cfg.ServiceSettings.EnableEmojiPicker,
-		"enable_gif_picker":                                       *cfg.ServiceSettings.EnableGifPicker,
-		"gfycat_api_key":                                          isDefault(*cfg.ServiceSettings.GfycatApiKey, model.SERVICE_SETTINGS_DEFAULT_GFYCAT_API_KEY),
-		"gfycat_api_secret":                                       isDefault(*cfg.ServiceSettings.GfycatApiSecret, model.SERVICE_SETTINGS_DEFAULT_GFYCAT_API_SECRET),
-		"experimental_enable_authentication_transfer":             *cfg.ServiceSettings.ExperimentalEnableAuthenticationTransfer,
-		"restrict_custom_emoji_creation":                          *cfg.ServiceSettings.RestrictCustomEmojiCreation,
-		"enable_testing":                                          cfg.ServiceSettings.EnableTesting,
-		"enable_developer":                                        *cfg.ServiceSettings.EnableDeveloper,
-		"enable_multifactor_authentication":                       *cfg.ServiceSettings.EnableMultifactorAuthentication,
-		"enforce_multifactor_authentication":                      *cfg.ServiceSettings.EnforceMultifactorAuthentication,
-		"enable_oauth_service_provider":                           cfg.ServiceSettings.EnableOAuthServiceProvider,
-		"connection_security":                                     *cfg.ServiceSettings.ConnectionSecurity,
-		"uses_letsencrypt":                                        *cfg.ServiceSettings.UseLetsEncrypt,
-		"forward_80_to_443":                                       *cfg.ServiceSettings.Forward80To443,
-		"maximum_login_attempts":                                  *cfg.ServiceSettings.MaximumLoginAttempts,
-		"session_length_web_in_days":                              *cfg.ServiceSettings.SessionLengthWebInDays,
-		"session_length_mobile_in_days":                           *cfg.ServiceSettings.SessionLengthMobileInDays,
-		"session_length_sso_in_days":                              *cfg.ServiceSettings.SessionLengthSSOInDays,
-		"session_cache_in_minutes":                                *cfg.ServiceSettings.SessionCacheInMinutes,
-		"session_idle_timeout_in_minutes":                         *cfg.ServiceSettings.SessionIdleTimeoutInMinutes,
-		"isdefault_site_url":                                      isDefault(*cfg.ServiceSettings.SiteURL, model.SERVICE_SETTINGS_DEFAULT_SITE_URL),
-		"isdefault_tls_cert_file":                                 isDefault(*cfg.ServiceSettings.TLSCertFile, model.SERVICE_SETTINGS_DEFAULT_TLS_CERT_FILE),
-		"isdefault_tls_key_file":                                  isDefault(*cfg.ServiceSettings.TLSKeyFile, model.SERVICE_SETTINGS_DEFAULT_TLS_KEY_FILE),
-		"isdefault_read_timeout":                                  isDefault(*cfg.ServiceSettings.ReadTimeout, model.SERVICE_SETTINGS_DEFAULT_READ_TIMEOUT),
-		"isdefault_write_timeout":                                 isDefault(*cfg.ServiceSettings.WriteTimeout, model.SERVICE_SETTINGS_DEFAULT_WRITE_TIMEOUT),
-		"isdefault_google_developer_key":                          isDefault(cfg.ServiceSettings.GoogleDeveloperKey, ""),
-		"isdefault_allow_cors_from":                               isDefault(*cfg.ServiceSettings.AllowCorsFrom, model.SERVICE_SETTINGS_DEFAULT_ALLOW_CORS_FROM),
-		"isdefault_cors_exposed_headers":                          isDefault(cfg.ServiceSettings.CorsExposedHeaders, ""),
-		"cors_allow_credentials":                                  *cfg.ServiceSettings.CorsAllowCredentials,
-		"cors_debug":                                              *cfg.ServiceSettings.CorsDebug,
+		"web_server_mode":                             *cfg.ServiceSettings.WebserverMode,
+		"enable_security_fix_alert":                   *cfg.ServiceSettings.EnableSecurityFixAlert,
+		"enable_insecure_outgoing_connections":        *cfg.ServiceSettings.EnableInsecureOutgoingConnections,
+		"enable_incoming_webhooks":                    cfg.ServiceSettings.EnableIncomingWebhooks,
+		"enable_outgoing_webhooks":                    cfg.ServiceSettings.EnableOutgoingWebhooks,
+		"enable_commands":                             *cfg.ServiceSettings.EnableCommands,
+		"enable_only_admin_integrations":              *cfg.ServiceSettings.EnableOnlyAdminIntegrations,
+		"enable_post_username_override":               cfg.ServiceSettings.EnablePostUsernameOverride,
+		"enable_post_icon_override":                   cfg.ServiceSettings.EnablePostIconOverride,
+		"enable_user_access_tokens":                   *cfg.ServiceSettings.EnableUserAccessTokens,
+		"enable_custom_emoji":                         *cfg.ServiceSettings.EnableCustomEmoji,
+		"enable_emoji_picker":                         *cfg.ServiceSettings.EnableEmojiPicker,
+		"enable_gif_picker":                           *cfg.ServiceSettings.EnableGifPicker,
+		"gfycat_api_key":                              isDefault(*cfg.ServiceSettings.GfycatApiKey, model.SERVICE_SETTINGS_DEFAULT_GFYCAT_API_KEY),
+		"gfycat_api_secret":                           isDefault(*cfg.ServiceSettings.GfycatApiSecret, model.SERVICE_SETTINGS_DEFAULT_GFYCAT_API_SECRET),
+		"experimental_enable_authentication_transfer": *cfg.ServiceSettings.ExperimentalEnableAuthenticationTransfer,
+		"restrict_custom_emoji_creation":              *cfg.ServiceSettings.RestrictCustomEmojiCreation,
+		"enable_testing":                              cfg.ServiceSettings.EnableTesting,
+		"enable_developer":                            *cfg.ServiceSettings.EnableDeveloper,
+		"enable_multifactor_authentication":           *cfg.ServiceSettings.EnableMultifactorAuthentication,
+		"enforce_multifactor_authentication":          *cfg.ServiceSettings.EnforceMultifactorAuthentication,
+		"enable_oauth_service_provider":               cfg.ServiceSettings.EnableOAuthServiceProvider,
+		"connection_security":                         *cfg.ServiceSettings.ConnectionSecurity,
+		"uses_letsencrypt":                            *cfg.ServiceSettings.UseLetsEncrypt,
+		"forward_80_to_443":                           *cfg.ServiceSettings.Forward80To443,
+		"maximum_login_attempts":                      *cfg.ServiceSettings.MaximumLoginAttempts,
+		"session_length_web_in_days":                  *cfg.ServiceSettings.SessionLengthWebInDays,
+		"session_length_mobile_in_days":               *cfg.ServiceSettings.SessionLengthMobileInDays,
+		"session_length_sso_in_days":                  *cfg.ServiceSettings.SessionLengthSSOInDays,
+		"session_cache_in_minutes":                    *cfg.ServiceSettings.SessionCacheInMinutes,
+		"session_idle_timeout_in_minutes":             *cfg.ServiceSettings.SessionIdleTimeoutInMinutes,
+		"isdefault_site_url":                          isDefault(*cfg.ServiceSettings.SiteURL, model.SERVICE_SETTINGS_DEFAULT_SITE_URL),
+		"isdefault_tls_cert_file":                     isDefault(*cfg.ServiceSettings.TLSCertFile, model.SERVICE_SETTINGS_DEFAULT_TLS_CERT_FILE),
+		"isdefault_tls_key_file":                      isDefault(*cfg.ServiceSettings.TLSKeyFile, model.SERVICE_SETTINGS_DEFAULT_TLS_KEY_FILE),
+		"isdefault_read_timeout":                      isDefault(*cfg.ServiceSettings.ReadTimeout, model.SERVICE_SETTINGS_DEFAULT_READ_TIMEOUT),
+		"isdefault_write_timeout":                     isDefault(*cfg.ServiceSettings.WriteTimeout, model.SERVICE_SETTINGS_DEFAULT_WRITE_TIMEOUT),
+		"isdefault_google_developer_key":              isDefault(cfg.ServiceSettings.GoogleDeveloperKey, ""),
+		"isdefault_allow_cors_from":                   isDefault(*cfg.ServiceSettings.AllowCorsFrom, model.SERVICE_SETTINGS_DEFAULT_ALLOW_CORS_FROM),
+		"isdefault_cors_exposed_headers":              isDefault(cfg.ServiceSettings.CorsExposedHeaders, ""),
+		"cors_allow_credentials":                      *cfg.ServiceSettings.CorsAllowCredentials,
+		"cors_debug":                                  *cfg.ServiceSettings.CorsDebug,
 		"isdefault_allowed_untrusted_internal_connections":        isDefault(*cfg.ServiceSettings.AllowedUntrustedInternalConnections, ""),
 		"restrict_post_delete":                                    *cfg.ServiceSettings.RestrictPostDelete,
 		"allow_edit_post":                                         *cfg.ServiceSettings.AllowEditPost,
@@ -308,15 +308,14 @@ func (a *App) trackConfig() {
 	})
 
 	a.SendDiagnostic(TRACK_CONFIG_SQL, map[string]interface{}{
-		"driver_name":                            *cfg.SqlSettings.DriverName,
-		"trace":                                  cfg.SqlSettings.Trace,
-		"max_idle_conns":                         *cfg.SqlSettings.MaxIdleConns,
-		"conn_max_lifetime_milliseconds":         *cfg.SqlSettings.ConnMaxLifetimeMilliseconds,
-		"max_open_conns":                         *cfg.SqlSettings.MaxOpenConns,
-		"data_source_replicas":                   len(cfg.SqlSettings.DataSourceReplicas),
-		"data_source_search_replicas":            len(cfg.SqlSettings.DataSourceSearchReplicas),
-		"query_timeout":                          *cfg.SqlSettings.QueryTimeout,
-		"enable_public_channels_materialization": *cfg.SqlSettings.EnablePublicChannelsMaterialization,
+		"driver_name":                    *cfg.SqlSettings.DriverName,
+		"trace":                          cfg.SqlSettings.Trace,
+		"max_idle_conns":                 *cfg.SqlSettings.MaxIdleConns,
+		"conn_max_lifetime_milliseconds": *cfg.SqlSettings.ConnMaxLifetimeMilliseconds,
+		"max_open_conns":                 *cfg.SqlSettings.MaxOpenConns,
+		"data_source_replicas":           len(cfg.SqlSettings.DataSourceReplicas),
+		"data_source_search_replicas":    len(cfg.SqlSettings.DataSourceSearchReplicas),
+		"query_timeout":                  *cfg.SqlSettings.QueryTimeout,
 	})
 
 	a.SendDiagnostic(TRACK_CONFIG_LOG, map[string]interface{}{
@@ -454,9 +453,9 @@ func (a *App) trackConfig() {
 	})
 
 	a.SendDiagnostic(TRACK_CONFIG_SAML, map[string]interface{}{
-		"enable":                              *cfg.SamlSettings.Enable,
-		"enable_sync_with_ldap":               *cfg.SamlSettings.EnableSyncWithLdap,
-		"enable_sync_with_ldap_include_auth":  *cfg.SamlSettings.EnableSyncWithLdapIncludeAuth,
+		"enable":                             *cfg.SamlSettings.Enable,
+		"enable_sync_with_ldap":              *cfg.SamlSettings.EnableSyncWithLdap,
+		"enable_sync_with_ldap_include_auth": *cfg.SamlSettings.EnableSyncWithLdapIncludeAuth,
 		"verify":                              *cfg.SamlSettings.Verify,
 		"encrypt":                             *cfg.SamlSettings.Encrypt,
 		"isdefault_scoping_idp_provider_id":   isDefault(*cfg.SamlSettings.ScopingIDPProviderId, ""),
