+++ model/config.go
@@ -691,17 +691,16 @@ type SSOSettings struct {
 }
 
 type SqlSettings struct {
-	DriverName                          *string
-	DataSource                          *string
-	DataSourceReplicas                  []string
-	DataSourceSearchReplicas            []string
-	MaxIdleConns                        *int
-	ConnMaxLifetimeMilliseconds         *int
-	MaxOpenConns                        *int
-	Trace                               bool
-	AtRestEncryptKey                    string
-	QueryTimeout                        *int
-	EnablePublicChannelsMaterialization *bool
+	DriverName                  *string
+	DataSource                  *string
+	DataSourceReplicas          []string
+	DataSourceSearchReplicas    []string
+	MaxIdleConns                *int
+	ConnMaxLifetimeMilliseconds *int
+	MaxOpenConns                *int
+	Trace                       bool
+	AtRestEncryptKey            string
+	QueryTimeout                *int
 }
 
 func (s *SqlSettings) SetDefaults() {
@@ -731,10 +730,6 @@ func (s *SqlSettings) SetDefaults() {
 
 	if s.QueryTimeout == nil {
 		s.QueryTimeout = NewInt(30)
-	}
-
-	if s.EnablePublicChannelsMaterialization == nil {
-		s.EnablePublicChannelsMaterialization = NewBool(true)
 	}
 }
