+++ store/storetest/mocks/ChannelStore.go
@@ -130,30 +130,6 @@ func (_m *ChannelStore) Delete(channelId string, time int64) store.StoreChannel 
 	return r0
 }
 
-// DisableExperimentalPublicChannelsMaterialization provides a mock function with given fields:
-func (_m *ChannelStore) DisableExperimentalPublicChannelsMaterialization() {
-	_m.Called()
-}
-
-// DropPublicChannels provides a mock function with given fields:
-func (_m *ChannelStore) DropPublicChannels() error {
-	ret := _m.Called()
-
-	var r0 error
-	if rf, ok := ret.Get(0).(func() error); ok {
-		r0 = rf()
-	} else {
-		r0 = ret.Error(0)
-	}
-
-	return r0
-}
-
-// EnableExperimentalPublicChannelsMaterialization provides a mock function with given fields:
-func (_m *ChannelStore) EnableExperimentalPublicChannelsMaterialization() {
-	_m.Called()
-}
-
 // Get provides a mock function with given fields: id, allowFromCache
 func (_m *ChannelStore) Get(id string, allowFromCache bool) store.StoreChannel {
 	ret := _m.Called(id, allowFromCache)
@@ -314,6 +290,22 @@ func (_m *ChannelStore) GetChannelMembersForExport(userId string, teamId string)
 	return r0
 }
 
+// GetChannelMembersTimezones provides a mock function with given fields: channelId
+func (_m *ChannelStore) GetChannelMembersTimezones(channelId string) store.StoreChannel {
+	ret := _m.Called(channelId)
+
+	var r0 store.StoreChannel
+	if rf, ok := ret.Get(0).(func(string) store.StoreChannel); ok {
+		r0 = rf(channelId)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(store.StoreChannel)
+		}
+	}
+
+	return r0
+}
+
 // GetChannelUnread provides a mock function with given fields: channelId, userId
 func (_m *ChannelStore) GetChannelUnread(channelId string, userId string) store.StoreChannel {
 	ret := _m.Called(channelId, userId)
@@ -433,21 +425,6 @@ func (_m *ChannelStore) GetMember(channelId string, userId string) store.StoreCh
 	var r0 store.StoreChannel
 	if rf, ok := ret.Get(0).(func(string, string) store.StoreChannel); ok {
 		r0 = rf(channelId, userId)
-	} else {
-		if ret.Get(0) != nil {
-			r0 = ret.Get(0).(store.StoreChannel)
-		}
-	}
-
-	return r0
-}
-
-func (_m *ChannelStore) GetChannelMembersTimezones(channelId string) store.StoreChannel {
-	ret := _m.Called(channelId)
-
-	var r0 store.StoreChannel
-	if rf, ok := ret.Get(0).(func(string) store.StoreChannel); ok {
-		r0 = rf(channelId)
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).(store.StoreChannel)
@@ -670,20 +647,6 @@ func (_m *ChannelStore) InvalidateChannelByName(teamId string, name string) {
 // InvalidateMemberCount provides a mock function with given fields: channelId
 func (_m *ChannelStore) InvalidateMemberCount(channelId string) {
 	_m.Called(channelId)
-}
-
-// IsExperimentalPublicChannelsMaterializationEnabled provides a mock function with given fields:
-func (_m *ChannelStore) IsExperimentalPublicChannelsMaterializationEnabled() bool {
-	ret := _m.Called()
-
-	var r0 bool
-	if rf, ok := ret.Get(0).(func() bool); ok {
-		r0 = rf()
-	} else {
-		r0 = ret.Get(0).(bool)
-	}
-
-	return r0
 }
 
 // IsUserInChannelUseCache provides a mock function with given fields: userId, channelId
